+++ NEWS.md
@@ -160,6 +160,11 @@ Deprecated or removed
   * `InexactError` now takes arguments: `InexactError(func::Symbol,
     type, -3)` now prints as `ERROR: InexactError: func(type, -3)`. ([#20005])
 
+  * The operating system identification functions: `is_linux`, `is_bsd`, `is_apple`, `is_unix`,
+    and `is_windows`, have been deprecated in favor of `Sys.islinux`, `Sys.isbsd`, `Sys.isapple`,
+    `Sys.isunix`, and `Sys.iswindows`, respectively ([#22182]).
+
+
 Julia v0.6.0 Release Notes
 ==========================
 
@@ -954,6 +959,7 @@ Command-line option changes
 [#22038]: https://github.com/JuliaLang/julia/issues/22038
 [#22062]: https://github.com/JuliaLang/julia/issues/22062
 [#22064]: https://github.com/JuliaLang/julia/issues/22064
+[#22182]: https://github.com/JuliaLang/julia/issues/22182
 [#22187]: https://github.com/JuliaLang/julia/issues/22187
 [#22188]: https://github.com/JuliaLang/julia/issues/22188
 [#22224]: https://github.com/JuliaLang/julia/issues/22224